requirements.txt
numpy==1.15.2
Pillow==5.3.0
scikit-image==0.15.0
opencv-python==3.4.2.16